Pick open port for DeviceTestContext early
If the caller doesn't specify a port, instead of
letting Tango select an open port, we get one just
before starting the `(Multi)DeviceTestContext`.
This way the port is known before starting,
and we can use it in an environment variable.

There is a tiny chance of race condition.  The OS
could allocate the port to another process between
object construction and when the Tango device is
actually launched.  The risk is considered very
low.  We have been using this approach for the
tests `test_event.py` and haven't seen failures
in CI due to that phenomenon.
